refactor(realm): extract module system from js-realm-shared
